nuking more stuff I don't understand
