+++ yarn-project/foundation/src/log/index.ts
@@ -1,6 +1,5 @@
 export * from './console.js';
-export * from './debug.js';
 export * from './pino-logger.js';
-export * from './log_history.js';
 export * from './log_fn.js';
 export * from './libp2p_logger.js';
+export * from './utils.js';
